+++ lib/sharezone_localizations/lib/sharezone_localizations.dart
@@ -1,5 +1,4 @@
 export 'localizations/sharezone_localizations.gen.dart';
-export 'src/app_locale_builder.dart';
-export 'src/app_locale_provider_bloc.dart';
+export 'src/app_locale_provider.dart';
 export 'src/app_locales.dart';
 export 'src/context_extension.dart';
